Convert frontend to typescript
