+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\langt\\HackGT\\testing\\virtual-checkin\\client\\src\\Example.js":"1"},{"size":1506,"mtime":1612487084493,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":6,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"1pw2zmj","C:\\Users\\langt\\HackGT\\testing\\virtual-checkin\\client\\src\\Example.js",["6","7","8","9","10","11"],{"ruleId":"12","severity":1,"message":"13","line":5,"column":13,"nodeType":"14","messageId":"15","endLine":5,"endColumn":15},{"ruleId":"12","severity":1,"message":"13","line":7,"column":20,"nodeType":"14","messageId":"15","endLine":7,"endColumn":22},{"ruleId":"12","severity":1,"message":"13","line":9,"column":20,"nodeType":"14","messageId":"15","endLine":9,"endColumn":22},{"ruleId":"12","severity":1,"message":"13","line":11,"column":20,"nodeType":"14","messageId":"15","endLine":11,"endColumn":22},{"ruleId":"12","severity":1,"message":"13","line":13,"column":20,"nodeType":"14","messageId":"15","endLine":13,"endColumn":22},{"ruleId":"12","severity":1,"message":"13","line":15,"column":20,"nodeType":"14","messageId":"15","endLine":15,"endColumn":22},"eqeqeq","Expected '===' and instead saw '=='.","BinaryExpression","unexpected"]
+[{"C:\\Users\\langt\\HackGT\\testing\\virtual-checkin\\client\\src\\Example.js":"1","C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\index.tsx":"2","C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\App.tsx":"3","C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\components\\YoutubeWrapper.tsx":"4"},{"size":1506,"mtime":1612487084493,"results":"5","hashOfConfig":"6"},{"size":623,"mtime":1613095830807,"results":"7","hashOfConfig":"8"},{"size":308,"mtime":1613098463429,"results":"9","hashOfConfig":"8"},{"size":1158,"mtime":1613097948324,"results":"10","hashOfConfig":"8"},{"filePath":"11","messages":"12","errorCount":0,"warningCount":6,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"1pw2zmj",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"15"},"s47dot",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":6,"fixableErrorCount":0,"fixableWarningCount":0,"source":"20","usedDeprecatedRules":"15"},"C:\\Users\\langt\\HackGT\\testing\\virtual-checkin\\client\\src\\Example.js",["21","22","23","24","25","26"],"C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\index.tsx",[],["27","28"],"C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\App.tsx",[],"C:\\Users\\langt\\HackGT\\virtual-checkin\\client\\src\\components\\YoutubeWrapper.tsx",["29","30","31","32","33","34"],"import React, { useState } from 'react';\r\nimport YouTube from 'react-youtube';\r\n\r\nfunction _get_state_text(state: any) {\r\n  if (state == 0) {\r\n    return \"ended\";\r\n  } else if (state == 1) {\r\n    return \"playing\"\r\n  } else if (state == 2) {\r\n    return \"paused\"\r\n  } else if (state == 3) {\r\n    return \"buffering\"\r\n  } else if (state == 5) {\r\n    return \"video cued\"\r\n  } else if (state == -1) {\r\n    return \"unstarted\"\r\n  } \r\n  return \"unknown\"\r\n}\r\n\r\n\r\nconst YoutubeWrapper: React.FC = () => {\r\n\r\n  const [isMuted, setIsMuted] = useState<String>('');\r\n  const [playerState, setPlayerState] = useState<String>('');\r\n\r\n  function _onChange(event: any) {\r\n    let player_state = _get_state_text(event.data)\r\n    let is_muted = String(event.target.isMuted())\r\n\r\n    setIsMuted(is_muted);\r\n    setPlayerState(player_state);\r\n  }\r\n\r\n  return (\r\n    <div>\r\n      <YouTube videoId=\"QH2-TGUlwu4\" opts={{height: '390', width: '640', playerVars: {autoplay: 1,}}} onStateChange={_onChange}/>\r\n      <div>\r\n        Player State: {playerState}\r\n      </div>\r\n      <div>\r\n        Is Muted: {isMuted}\r\n      </div>\r\n    </div>\r\n  )\r\n  }\r\n\r\nexport default YoutubeWrapper;\r\n",{"ruleId":"35","severity":1,"message":"36","line":5,"column":13,"nodeType":"37","messageId":"38","endLine":5,"endColumn":15},{"ruleId":"35","severity":1,"message":"36","line":7,"column":20,"nodeType":"37","messageId":"38","endLine":7,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":9,"column":20,"nodeType":"37","messageId":"38","endLine":9,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":11,"column":20,"nodeType":"37","messageId":"38","endLine":11,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":13,"column":20,"nodeType":"37","messageId":"38","endLine":13,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":15,"column":20,"nodeType":"37","messageId":"38","endLine":15,"endColumn":22},{"ruleId":"39","replacedBy":"40"},{"ruleId":"41","replacedBy":"42"},{"ruleId":"35","severity":1,"message":"36","line":5,"column":13,"nodeType":"37","messageId":"38","endLine":5,"endColumn":15},{"ruleId":"35","severity":1,"message":"36","line":7,"column":20,"nodeType":"37","messageId":"38","endLine":7,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":9,"column":20,"nodeType":"37","messageId":"38","endLine":9,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":11,"column":20,"nodeType":"37","messageId":"38","endLine":11,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":13,"column":20,"nodeType":"37","messageId":"38","endLine":13,"endColumn":22},{"ruleId":"35","severity":1,"message":"36","line":15,"column":20,"nodeType":"37","messageId":"38","endLine":15,"endColumn":22},"eqeqeq","Expected '===' and instead saw '=='.","BinaryExpression","unexpected","no-native-reassign",["43"],"no-negated-in-lhs",["44"],"no-global-assign","no-unsafe-negation"]
