--- client/src/App.js
@@ -1,16 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-import Example from './Example.js'
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <Example></Example>
-      </header>
-    </div>
-  );
-}
-
-export default App;
