+++ client/src/App.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import './App.css';
+
+// components
+import YoutubeWrapper from './components/YoutubeWrapper';
+
+const App: React.FC = () => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <YoutubeWrapper/>
+      </header>
+    </div>
+  );
+};
+
+export default App;
